backend/list_models.py
import os
from google import genai
from dotenv import load_dotenv

load_dotenv()

api_key = os.getenv("GEMINI_API_KEY") or os.getenv("GOOGLE_API_KEY")
if not api_key:
    print("No API Key found")
else:
    client = genai.Client(api_key=api_key)
    try:
        for m in client.models.list():
            print(m.name)
    except Exception as e:
        print(f"Error: {e}")
