backend/requirements.txt
fastapython-dotenv
uvicorn
google-genai
fastapi
sqlalchemy
elevenlabs
edge-tts
